Actualiza esquema para aceptar decimales para el valor del tiempo

+++ schemas/scores.js
@@ -1,10 +1,10 @@
-import z from 'zod';
+import z from "zod";
 
 const scoreSchema = z.object({
     palabras_acertadas: z.number().int().positive(),
     palabras_totales: z.number().int().positive(),
-    tiempo_segundos: z.number().int().positive(),
-    precision_porcentaje: z.number().positive().min(0).max(100)
+    tiempo_segundos: z.number().positive(),
+    precision_porcentaje: z.number().positive().min(0).max(100),
 });
 
 export function validateScore(object) {
